Move gdal/Implicits into methods on their classes
